# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type NetworkCategory = typeof NetworkCategory[keyof typeof NetworkCategory];


export const NetworkCategory = {
  VLAN: 'VLAN',
  Gateway: 'Gateway',
  DHCP: 'DHCP',
  DNS: 'DNS',
  Routing: 'Routing',
  ACL: 'ACL',
  NAT: 'NAT',
  Wireless: 'Wireless',
  Other: 'Other',
} as const;

export type ReviewStatus = typeof ReviewStatus[keyof typeof ReviewStatus];


export const ReviewStatus = {
  pending: 'pending',
  accepted: 'accepted',
  edited: 'edited',
  rejected: 'rejected',
} as const;

export type Severity = typeof Severity[keyof typeof Severity];


export const Severity = {
  Low: 'Low',
  Medium: 'Medium',
  High: 'High',
} as const;

export type OsiLayer = typeof OsiLayer[keyof typeof OsiLayer];


export const OsiLayer = {
  Layer_1: 'Layer 1',
  Layer_2: 'Layer 2',
  Layer_3: 'Layer 3',
  Layer_4: 'Layer 4',
  Layer_7: 'Layer 7',
  Unknown: 'Unknown',
} as const;

export interface CheckerFinding {
  ruleId: string;
  category: NetworkCategory;
  finding: string;
  severity: Severity;
  evidence: string[];
}

export interface Diagnosis {
  rootCause: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence: number;
  osiLayer: OsiLayer;
  evidence: string[];
  nextCommand: string;
  fixSteps: string[];
  risk: Severity;
  humanReviewRequired: boolean;
}

export interface ReviewEdits {
  rootCause?: string;
  nextCommand?: string;
  fixSteps?: string[];
}

export interface AnalyzeInput {
  category: NetworkCategory;
  /**
     * @minLength 1
     * @maxLength 5000
     */
  symptom: string;
  /** @maxLength 20000 */
  output: string;
}

export interface AnalyzeResponse {
  id: number;
  category: NetworkCategory;
  checks: CheckerFinding[];
  diagnosis: Diagnosis;
  createdAt: string;
}

export interface DiagnosisSummary {
  id: number;
  category: NetworkCategory;
  rootCause: string;
  confidence: number;
  severity: Severity;
  reviewStatus: ReviewStatus;
  createdAt: string;
}

export type DiagnosisRecord = DiagnosisSummary & ({
  symptom: string;
  output: string;
  checks: CheckerFinding[];
  diagnosis: Diagnosis;
  /** @nullable */
  reviewNotes: string | null;
  reviewEdits: ReviewEdits | null;
});

export type ReviewInputStatus = typeof ReviewInputStatus[keyof typeof ReviewInputStatus];


export const ReviewInputStatus = {
  accepted: 'accepted',
  edited: 'edited',
  rejected: 'rejected',
} as const;

export interface ReviewInput {
  status: ReviewInputStatus;
  /** @maxLength 2000 */
  notes?: string;
  edits?: ReviewEdits | null;
}

export interface CategoryCount {
  category: NetworkCategory;
  count: number;
}

export type DashboardMetricsSeverityCounts = {
  high: number;
  medium: number;
  low: number;
};

export type DashboardMetricsReviewCounts = {
  pending: number;
  accepted: number;
  edited: number;
  rejected: number;
};

export interface DashboardMetrics {
  totalDiagnoses: number;
  severityCounts: DashboardMetricsSeverityCounts;
  /** @nullable */
  averageConfidence: number | null;
  reviewCounts: DashboardMetricsReviewCounts;
  categoryCounts: CategoryCount[];
  dataAvailability: string;
}

export interface Case {
  caseId: string;
  category: NetworkCategory;
  symptom: string;
  command: string;
  observedOutput: string;
  expectedFault: string;
  osiLayer: OsiLayer;
  concept: string;
  severity: Severity;
}

export interface ErrorResponse {
  error: string;
}

export type LimitParameter = number;

export type CategoryParameter = NetworkCategory;

export type CaseCategoryParameter = NetworkCategory;

export type ReviewStatusParameter = ReviewStatus;

export type ListDiagnosesParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: LimitParameter;
category?: CategoryParameter;
reviewStatus?: ReviewStatusParameter;
};

export type ListCasesParams = {
category?: CaseCategoryParameter;
};