# lib/api-zod/src/generated/types/analyzeInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NetworkCategory } from './networkCategory';

export interface AnalyzeInput {
  category: NetworkCategory;
  /**
     * @minLength 1
     * @maxLength 5000
     */
  symptom: string;
  /** @maxLength 20000 */
  output: string;
}