# lib/api-zod/src/generated/types/case.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NetworkCategory } from './networkCategory';
import type { OsiLayer } from './osiLayer';
import type { Severity } from './severity';

export interface Case {
  caseId: string;
  category: NetworkCategory;
  symptom: string;
  command: string;
  observedOutput: string;
  expectedFault: string;
  osiLayer: OsiLayer;
  concept: string;
  severity: Severity;
}