# lib/api-zod/src/generated/types/categoryCount.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NetworkCategory } from './networkCategory';

export interface CategoryCount {
  category: NetworkCategory;
  count: number;
}