# lib/api-zod/src/generated/types/diagnosisRecord.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CheckerFinding } from './checkerFinding';
import type { Diagnosis } from './diagnosis';
import type { DiagnosisSummary } from './diagnosisSummary';
import type { ReviewEdits } from './reviewEdits';

export type DiagnosisRecord = DiagnosisSummary & ({
  symptom: string;
  output: string;
  checks: CheckerFinding[];
  diagnosis: Diagnosis;
  /** @nullable */
  reviewNotes: string | null;
  reviewEdits: ReviewEdits | null;
});