# lib/api-zod/src/generated/types/networkCategory.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type NetworkCategory = typeof NetworkCategory[keyof typeof NetworkCategory];


export const NetworkCategory = {
  VLAN: 'VLAN',
  Gateway: 'Gateway',
  DHCP: 'DHCP',
  DNS: 'DNS',
  Routing: 'Routing',
  ACL: 'ACL',
  NAT: 'NAT',
  Wireless: 'Wireless',
  Other: 'Other',
} as const;