# lib/api-zod/src/generated/types/osiLayer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type OsiLayer = typeof OsiLayer[keyof typeof OsiLayer];


export const OsiLayer = {
  Layer_1: 'Layer 1',
  Layer_2: 'Layer 2',
  Layer_3: 'Layer 3',
  Layer_4: 'Layer 4',
  Layer_7: 'Layer 7',
  Unknown: 'Unknown',
} as const;