# lib/api-zod/src/generated/types/reviewEdits.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReviewEdits {
  rootCause?: string;
  nextCommand?: string;
  fixSteps?: string[];
}