# lib/api-zod/src/generated/types/reviewStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type ReviewStatus = typeof ReviewStatus[keyof typeof ReviewStatus];


export const ReviewStatus = {
  pending: 'pending',
  accepted: 'accepted',
  edited: 'edited',
  rejected: 'rejected',
} as const;