# lib/api-zod/src/generated/types/severity.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type Severity = typeof Severity[keyof typeof Severity];


export const Severity = {
  Low: 'Low',
  Medium: 'Medium',
  High: 'High',
} as const;